requirements.txt
Flask==0.12.1
Jinja2==2.9.6
MarkupSafe==1.0
Werkzeug==0.12.1
click==6.7
itsdangerous==0.24
pyephem==3.7.6.0
